Referral demo

+++ server/src/internal/products/product-items/productItemInitUtils.ts
@@ -1,12 +1,11 @@
 import {
-  EntInterval,
+  AppEnv,
   Entitlement,
-  ErrCode,
   Feature,
   Price,
   Product,
   ProductItem,
-  ProductItemBehavior,
+  UsageModel,
 } from "@autumn/shared";
 import { itemToPriceAndEnt } from "./mapFromItem.js";
 
@@ -15,6 +14,8 @@ import { EntitlementService } from "../entitlements/EntitlementService.js";
 import { SupabaseClient } from "@supabase/supabase-js";
 
 import { validateProductItems } from "./validateProductItems.js";
+import { createFeaturesFromItems } from "./createFeaturesFromItems.js";
+import { FeatureService } from "@/internal/features/FeatureService.js";
 
 const isNewItem = (item: ProductItem) => {
   return !item.entitlement_id && !item.price_id;
@@ -65,10 +66,48 @@ const updateDbPricesAndEnts = async ({
     }),
   ]);
 
-  await EntitlementService.deleteByIds({
+  // Check if any custom prices use this entitlement...
+  let deletedEntIds = deletedEnts.map((ent) => ent.id!);
+  let customPrices = await PriceService.getInIds({
     sb,
-    entitlementIds: deletedEnts.map((ent) => ent.id!),
+    entitlementIds: deletedEntIds,
   });
+
+  if (customPrices.length == 0) {
+    // Update the entitlement to be custom...
+    await EntitlementService.deleteByIds({
+      sb,
+      entitlementIds: deletedEntIds,
+    });
+  } else {
+    let updateOrDelete: any = [];
+    for (const ent of deletedEnts) {
+      let hasCustomPrice = customPrices.some(
+        (price) => price.entitlement_id == ent.id
+      );
+
+      if (hasCustomPrice) {
+        updateOrDelete.push(
+          EntitlementService.update({
+            sb,
+            entitlementId: ent.id!,
+            updates: {
+              is_custom: true,
+            },
+          })
+        );
+      } else {
+        updateOrDelete.push(
+          EntitlementService.deleteByIds({
+            sb,
+            entitlementIds: [ent.id!],
+          })
+        );
+      }
+    }
+
+    await Promise.all(updateOrDelete);
+  }
 };
 
 const handleCustomProductItems = async ({
@@ -130,6 +169,8 @@ export const handleNewProductItems = async ({
   isCustom: boolean;
   newVersion?: boolean;
 }) => {
+  // Create features if not exist...
+
   if (!newItems) {
     return {
       prices: [],
@@ -138,10 +179,15 @@ export const handleNewProductItems = async ({
   }
 
   // Validate product items...
-  validateProductItems({
+
+  let { allFeatures, newFeatures } = validateProductItems({
     newItems,
     features,
+    orgId: product.org_id!,
+    env: product.env as AppEnv,
   });
+
+  features = allFeatures;
 
   let newPrices: Price[] = [];
   let newEnts: Entitlement[] = [];
@@ -210,12 +256,11 @@ export const handleNewProductItems = async ({
     `Ents: new(${newEnts.length}), updated(${updatedEnts.length}), deleted(${deletedEnts.length})`
   );
 
-  // console.log("New prices: ", newPrices);
-  // throw new Error("test");
-  // console.log("Updated prices: ", updatedPrices);
-  // console.log("New entitlements: ", newEnts);
-  // console.log("Updated entitlements: ", updatedEnts);
-  // throw new Error("test");
+  // Create new features
+  await FeatureService.insert({
+    sb,
+    data: newFeatures,
+  });
 
   if (isCustom || newVersion) {
     return handleCustomProductItems({
